test: speed up per-commit test checks
