+++ test/domain/services/tmux_service_control_mode_test.dart
@@ -1009,9 +1009,22 @@ void main() {
         final client = _MockSshClient();
         final session = _buildSession(client);
         const service = TmuxService();
+        final stdoutController = StreamController<Uint8List>();
+        final controlSession = _buildInteractiveExecSession(
+          stdoutController: stdoutController,
+          onWrite: (value) {
+            if (value.startsWith('refresh-client ')) {
+              scheduleMicrotask(
+                () => stdoutController.add(
+                  _utf8Bytes('%begin 1 1 0\n%end 1 1 0\n'),
+                ),
+              );
+            }
+          },
+        );
         final execSessions = Queue<SSHSession>.from([
           _buildOpenExecSession(stdout: 'zsh\n/usr/bin/tmux\n${_doneMarker()}'),
-          _buildOpenExecSession(),
+          controlSession,
         ]);
 
         when(
@@ -1031,6 +1044,7 @@ void main() {
             pty: any(named: 'pty'),
           ),
         );
+        await untilCalled(() => controlSession.write(any()));
 
         final command =
             verify(
@@ -1050,6 +1064,7 @@ void main() {
         expect(command, isNot(contains('set status off')));
 
         await subscription.cancel();
+        await stdoutController.close();
       },
     );
