+++ test/widget/home_screen_test.dart
@@ -7,7 +7,6 @@ import 'package:drift/native.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:go_router/go_router.dart';
 import 'package:mocktail/mocktail.dart';
 
 import 'package:monkeyssh/data/database/database.dart';
@@ -206,70 +205,10 @@ ActiveConnection _buildActiveConnection({
   ),
 );
 
-// These tests are skipped because the HomeScreen uses StreamProviders
-// which don't settle in widget tests (continuous database watchers).
-// The underlying repository and service tests pass (127 tests).
 void main() {
   setUpAll(() {
     registerFallbackValue(<int>[]);
   });
-
-  Widget buildTestWidget(AppDatabase db, {Size size = const Size(800, 600)}) =>
-      ProviderScope(
-        overrides: [
-          databaseProvider.overrideWithValue(db),
-          transferIntentServiceProvider.overrideWith(
-            (ref) => _TestTransferIntentService(),
-          ),
-          homeScreenShortcutServiceProvider.overrideWith(
-            (ref) => _TestHomeScreenShortcutService(),
-          ),
-          pinnedHomeScreenShortcutHostIdsProvider.overrideWith(
-            (ref) => Stream<Set<int>>.value(const <int>{}),
-          ),
-        ],
-        child: MediaQuery(
-          data: MediaQueryData(size: size),
-          child: MaterialApp.router(
-            routerConfig: GoRouter(
-              routes: [
-                GoRoute(
-                  path: '/',
-                  builder: (context, state) => const HomeScreen(),
-                ),
-                GoRoute(
-                  path: '/settings',
-                  builder: (context, state) => const Scaffold(),
-                ),
-                GoRoute(
-                  path: '/hosts/add',
-                  builder: (context, state) => const Scaffold(),
-                ),
-                GoRoute(
-                  path: '/hosts/:id/edit',
-                  builder: (context, state) => const Scaffold(),
-                ),
-                GoRoute(
-                  path: '/terminal/:hostId',
-                  builder: (context, state) => const Scaffold(),
-                ),
-                GoRoute(
-                  path: '/keys/add',
-                  builder: (context, state) => const Scaffold(),
-                ),
-                GoRoute(
-                  path: '/snippets',
-                  builder: (context, state) => const Scaffold(),
-                ),
-                GoRoute(
-                  path: '/port-forwards',
-                  builder: (context, state) => const Scaffold(),
-                ),
-              ],
-            ),
-          ),
-        ),
-      );
 
   Widget buildMobileHomeScreen({
     required AppDatabase db,
@@ -293,99 +232,6 @@ void main() {
       data: MediaQueryData(size: size),
       child: const MaterialApp(home: HomeScreen()),
     ),
-  );
-
-  // Skip tests due to StreamProvider not settling in widget tests
-  group(
-    'HomeScreen Desktop Layout',
-    skip: true, // StreamProvider tests hang - use integration tests instead
-    () {
-      testWidgets('displays app title in sidebar', (tester) async {
-        final db = AppDatabase.forTesting(NativeDatabase.memory());
-        addTearDown(db.close);
-
-        await tester.pumpWidget(buildTestWidget(db));
-        await tester.pump();
-        await tester.pump(const Duration(milliseconds: 100));
-
-        expect(find.text('Flutty'), findsOneWidget);
-      });
-
-      testWidgets('displays navigation items in sidebar', (tester) async {
-        final db = AppDatabase.forTesting(NativeDatabase.memory());
-        addTearDown(db.close);
-
-        await tester.pumpWidget(buildTestWidget(db));
-        await tester.pump();
-        await tester.pump(const Duration(milliseconds: 100));
-
-        expect(find.text('Hosts'), findsOneWidget);
-        expect(find.text('Keys'), findsOneWidget);
-        expect(find.text('Snippets'), findsOneWidget);
-        expect(find.text('Port Forwarding'), findsOneWidget);
-        expect(find.text('Settings'), findsOneWidget);
-      });
-
-      testWidgets('displays settings icon', (tester) async {
-        final db = AppDatabase.forTesting(NativeDatabase.memory());
-        addTearDown(db.close);
-
-        await tester.pumpWidget(buildTestWidget(db));
-        await tester.pump();
-        await tester.pump(const Duration(milliseconds: 100));
-
-        expect(find.byIcon(Icons.settings_outlined), findsOneWidget);
-      });
-
-      testWidgets('displays sidebar navigation icons', (tester) async {
-        final db = AppDatabase.forTesting(NativeDatabase.memory());
-        addTearDown(db.close);
-
-        await tester.pumpWidget(buildTestWidget(db));
-        await tester.pump();
-        await tester.pump(const Duration(milliseconds: 100));
-
-        expect(find.byIcon(Icons.dns_rounded), findsOneWidget);
-        expect(find.byIcon(Icons.key_rounded), findsOneWidget);
-        expect(find.byIcon(Icons.code_rounded), findsOneWidget);
-        expect(find.byIcon(Icons.swap_horiz_rounded), findsOneWidget);
-      });
-    },
-  );
-
-  group(
-    'HomeScreen Mobile Layout',
-    skip: true, // StreamProvider tests hang - use integration tests instead
-    () {
-      testWidgets('displays bottom navigation bar on narrow screens', (
-        tester,
-      ) async {
-        final db = AppDatabase.forTesting(NativeDatabase.memory());
-        addTearDown(db.close);
-
-        await tester.pumpWidget(
-          buildTestWidget(db, size: const Size(400, 800)),
-        );
-        await tester.pump();
-        await tester.pump(const Duration(milliseconds: 100));
-
-        expect(find.byType(NavigationBar), findsOneWidget);
-      });
-
-      testWidgets('displays app bar on narrow screens', (tester) async {
-        final db = AppDatabase.forTesting(NativeDatabase.memory());
-        addTearDown(db.close);
-
-        await tester.pumpWidget(
-          buildTestWidget(db, size: const Size(400, 800)),
-        );
-        await tester.pump();
-        await tester.pump(const Duration(milliseconds: 100));
-
-        expect(find.byType(AppBar), findsOneWidget);
-        expect(find.text('Flutty'), findsOneWidget);
-      });
-    },
   );
 
   group('HomeScreen reorder affordance', () {
