+++ test/widget/keys_screen_test.dart
@@ -1,6 +1,5 @@
 // ignore_for_file: public_member_api_docs
 
-import 'package:drift/native.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -8,10 +7,6 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:monkeyssh/data/database/database.dart';
 import 'package:monkeyssh/presentation/providers/entity_list_providers.dart';
 import 'package:monkeyssh/presentation/screens/keys_screen.dart';
-
-// Most KeysScreen tests are skipped because the screen uses StreamProviders
-// which don't settle in widget tests (continuous database watchers).
-// The underlying repository tests provide coverage.
 
 final _testKey = SshKey(
   id: 1,
@@ -54,24 +49,6 @@ void main() {
       expect(find.text('Generate Key'), findsOneWidget);
       expect(find.text('Import Key'), findsOneWidget);
     });
-
-    testWidgets(
-      'shows loading indicator initially',
-      (tester) async {
-        final db = AppDatabase.forTesting(NativeDatabase.memory());
-        addTearDown(db.close);
-
-        await tester.pumpWidget(
-          ProviderScope(
-            overrides: [databaseProvider.overrideWithValue(db)],
-            child: const MaterialApp(home: KeysScreen()),
-          ),
-        );
-
-        expect(find.byType(CircularProgressIndicator), findsOneWidget);
-      },
-      skip: true, // Drift StreamProviders leave pending timers in test cleanup
-    );
   });
 
   group('private key clipboard gating', () {
