--- test/widget/terminal_tabs_test.dart
@@ -1,64 +1,0 @@
-// ignore_for_file: public_member_api_docs
-
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter_test/flutter_test.dart';
-
-import 'package:monkeyssh/presentation/widgets/terminal_tabs.dart';
-
-// Some TerminalTabBar tests are skipped due to framework issues when running
-// in parallel with other tests.
-void main() {
-  group(
-    'TerminalTabBar',
-    skip: true, // Flaky when run with other tests
-    () {
-      testWidgets('does not render when no sessions', (tester) async {
-        await tester.pumpWidget(
-          ProviderScope(
-            child: MaterialApp(
-              home: Scaffold(
-                body: TerminalTabBar(
-                  activeHostId: null,
-                  onTabSelected: (_) {},
-                  onTabClosed: (_) {},
-                  onNewTab: () {},
-                ),
-              ),
-            ),
-          ),
-        );
-
-        // Should render as empty SizedBox
-        expect(find.byType(ListView), findsNothing);
-      });
-
-      testWidgets('renders as SizedBox.shrink when empty', (tester) async {
-        await tester.pumpWidget(
-          ProviderScope(
-            child: MaterialApp(
-              home: Scaffold(
-                body: TerminalTabBar(
-                  activeHostId: null,
-                  onTabSelected: (_) {},
-                  onTabClosed: (_) {},
-                  onNewTab: () {},
-                ),
-              ),
-            ),
-          ),
-        );
-
-        // Add button should not be visible (inside shrunk box)
-        expect(find.byType(SizedBox), findsWidgets);
-      });
-    },
-  );
-
-  group('SshConnectionState', () {
-    test('enum values exist', () {
-      // Verify the enum values we expect
-      expect(true, isTrue);
-    });
-  });
-}
